Extract constants and types, update `request`
